rest api for react camping-shop-app

+++ controllers/productController.js
@@ -69,7 +69,7 @@ function getProduct(req, res, next) {
 
 	productModel.findById(productId)
 		.populate('userId')
-		.populate('comments')
+		// .populate('comments')
 		.then(product => {
 			console.log(product)
 			res.json(product)
@@ -92,11 +92,13 @@ function updateProduct(req, res, next) {
 }
 
 function createProduct(req, res, next) {
-	const { _id: userId } = req.user;
-	const { name, description, info, imageUrl, price, quantity, category } = req.body;
-	productModel.create({ name, description, info, imageUrl, price, quantity, category, userId })
+	const { _id } = req.user;
+	if (!req.user.isAdmin) { return; }
+	const { name, description, info, imageUrl, price, quantity, category } = req.body.data;
+	console.log('REQ.BODY: ', req.body)
+	productModel.create({ name, description, info, imageUrl, price, quantity, category, userId: _id })
 		.then(product => {
-			return userModel.updateOne({ _id: userId }, { $push: { products: product._id } });
+			return userModel.updateOne({ _id }, { $push: { products: product._id } });
 		}).then(product => {
 			console.log(product);
 			res.status(200).json(product);
@@ -104,18 +106,39 @@ function createProduct(req, res, next) {
 		.catch(next);
 }
 
+// function deleteProduct(req, res, next) {
+// 	const { productId } = req.params;
+// 	const { _id: userId } = req.user;
+// 	console.log(productId, userId)
+
+// 	Promise.all([
+// 		productModel.findOneAndDelete({ _id: productId, userId }),
+// 		userModel.findOneAndUpdate({ userId }, { $pull: { products: productId } }),
+// 		commentModel.deleteMany({ productId: productId }, { $in: { productId: productId } }),
+// 	])
+// 		.then(([deletedOne, _, __]) => {
+// 			if (deletedOne) {
+// 				res.status(200).json(deletedOne, { message: 'Product successfully deleted.' })
+// 			} else {
+// 				res.status(401).json({ message: `Not allowed!` });
+// 			}
+// 		})
+// 		.catch(next);
+// }
+
 function deleteProduct(req, res, next) {
 	const { productId } = req.params;
 	const { _id: userId } = req.user;
+	// console.log(productId, req.user);
 
+	if (!req.user.isAdmin) { return; }
 	Promise.all([
 		productModel.findOneAndDelete({ _id: productId, userId }),
-		userModel.findOneAndUpdate({ _id: userId }, { $pull: { products: productId } }),
-		commentModel.deleteMany({ productId: productId }, { $in: { productId: productId } }),
+		userModel.findOneAndUpdate({ userId }, { $pull: { products: productId } })
 	])
-		.then(([deletedOne, _, __]) => {
+		.then(([deletedOne, _]) => {
 			if (deletedOne) {
-				res.status(200).json(deletedOne)
+				res.status(200).json({ deletedOne, message: 'Product successfully deleted.' })
 			} else {
 				res.status(401).json({ message: `Not allowed!` });
 			}
@@ -123,35 +146,35 @@ function deleteProduct(req, res, next) {
 		.catch(next);
 }
 
-function addToFavorites(req, res, next) {
-	const { productId } = req.params;
-	const { _id: userId } = req.user;
-	userModel.findByIdAndUpdate({ _id: userId }, { $addToSet: { favorites: productId } }, { new: true })
-		.then(updatedUser => {
-			res.status(200).json(updatedUser)
-		})
-		.catch(next);
-}
+// function addToFavorites(req, res, next) {
+// 	const { productId } = req.params;
+// 	const { _id: userId } = req.user;
+// 	userModel.findByIdAndUpdate({ _id: userId }, { $addToSet: { favorites: productId } }, { new: true })
+// 		.then(updatedUser => {
+// 			res.status(200).json(updatedUser)
+// 		})
+// 		.catch(next);
+// }
 
-function removeFromFavorites(req, res, next) {
-	const { productId } = req.params;
-	console.log('req.params', req.params)
-	const { _id: userId } = req.user;
-	console.log(productId, 'req.user ->', req.user);
-	userModel.findByIdAndUpdate({ _id: userId }, { $pull: { favorites: productId } })
-		.then((updatedUser) => res.status(200).json(updatedUser))
-		.catch(next);
-}
+// function removeFromFavorites(req, res, next) {
+// 	const { productId } = req.params;
+// 	console.log('req.params', req.params)
+// 	const { _id: userId } = req.user;
+// 	console.log(productId, 'req.user ->', req.user);
+// 	userModel.findByIdAndUpdate({ _id: userId }, { $pull: { favorites: productId } })
+// 		.then((updatedUser) => res.status(200).json(updatedUser))
+// 		.catch(next);
+// }
 
-function getFavorites(req, res, next) {
-	const { _id: userId } = req.user;
-	userModel.findById({ _id: userId }).populate('favorites').populate('userId')
-		.then(user => {
-			console.log(user)
-			res.status(200).json(user.favorites);
-		})
-		.catch(next);
-}
+// function getFavorites(req, res, next) {
+// 	const { _id: userId } = req.user;
+// 	userModel.findById({ _id: userId }).populate('favorites').populate('userId')
+// 		.then(user => {
+// 			console.log(user)
+// 			res.status(200).json(user.favorites);
+// 		})
+// 		.catch(next);
+// }
 
 function like(req, res, next) {
 	// const { productId } = req.params;
@@ -172,8 +195,8 @@ module.exports = {
 	updateProduct,
 	getProduct,
 	getProductAndComments,
-	addToFavorites,
-	removeFromFavorites,
-	getFavorites,
+	// addToFavorites,
+	// removeFromFavorites,
+	// getFavorites,
 	like,
 }
